+++ src/main/java/space/efremov/otusspringlibrary/dao/jdbc/EntityToMapConverter.java
@@ -1,8 +1,8 @@
 package space.efremov.otusspringlibrary.dao.jdbc;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.stereotype.Service;
 import space.efremov.otusspringlibrary.domain.AbstractEntity;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -16,7 +16,7 @@ public class EntityToMapConverter {
         return oMapper.convertValue(entity, Map.class);
     }
 
-    public Map<String, Object> getIdParam(Integer id) {
+    public Map<String, Object> getIdParam(Long id) {
         final HashMap<String, Object> params = new HashMap<>(1);
         params.put("id", id);
         return params;
