+++ src/main/java/space/efremov/otusspringlibrary/dao/jdbc/PublishDao.java
@@ -16,7 +16,7 @@ import java.util.stream.Collectors;
 public class PublishDao extends AbstractJdbcDao<Publisher> {
 
     private final RowMapper<Publisher> rowMapper = (RowMapper<Publisher>) (resultSet, i) -> {
-        Integer id = resultSet.getInt("id");
+        Long id = resultSet.getLong("id");
         String name = resultSet.getString("name");
         return new Publisher(id, name);
     };
@@ -26,8 +26,8 @@ public class PublishDao extends AbstractJdbcDao<Publisher> {
     }
 
     @Override
-    public Integer count() {
-        return jdbc.queryForObject("select count(*) from publisher", Collections.EMPTY_MAP, Integer.class);
+    public Long count() {
+        return jdbc.queryForObject("select count(*) from publisher", Collections.EMPTY_MAP, Long.class);
     }
 
     @Override
@@ -37,7 +37,7 @@ public class PublishDao extends AbstractJdbcDao<Publisher> {
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Long id) {
         getById(id);
         final Map<String, Object> idParam = converter.getIdParam(id);
         jdbc.update("delete from book where publisher_id = :id", idParam);
@@ -45,7 +45,7 @@ public class PublishDao extends AbstractJdbcDao<Publisher> {
     }
 
     @Override
-    public Publisher getById(Integer id) {
+    public Publisher getById(Long id) {
         try {
             return jdbc.queryForObject("select * from publisher where id = :id", converter.getIdParam(id), rowMapper);
         } catch (EmptyResultDataAccessException e) {
@@ -56,6 +56,6 @@ public class PublishDao extends AbstractJdbcDao<Publisher> {
     @Override
     public List<Publisher> getAll() {
         final List<Map> rows = jdbc.queryForList("select * from publisher", Collections.EMPTY_MAP);
-        return rows.stream().map(m -> new Publisher((Integer) m.get("id"), (String) m.get("name"))).collect(Collectors.toList());
+        return rows.stream().map(m -> new Publisher((Long) m.get("id"), (String) m.get("name"))).collect(Collectors.toList());
     }
 }
